browser fixes
fixed the code to work for FF and Edge. IE is still a major pain to deal
with and I haven't had very much success trying to get the IE debugger
to work properly. fixed the image rows and full page rows to be more
responsive as the calendar.byu.edu/calendar-widget-demo was having some
problems in that area. the scss for the image-rows was having problems
in FF and Edge with the following selector ":host([type="image"]) so I
had to update it to ":host([image-row]) and updated the component to
match it.